Remove deprecated API middleware, OAuth handling, routing, server management, token management, and user profile management code
- Deleted middleware.go, oauth.go, routes.go, server.go, tokens.go, user_profiles.go, and users.go files.
- This cleanup removes unused code and simplifies the codebase.
