+++ cmd/api/RESTRUCTURING_GUIDE.md
@@ -1,0 +1,135 @@
+# API Restructuring - New Folder Structure
+
+## Overview
+This document outlines the restructured Go REST API following Go best practices for better organization, maintainability, and separation of concerns.
+
+## New Structure
+
+```
+cmd/api/
+├── main_new.go          # New entry point (replace main.go)
+├── app/                 # Application core
+│   ├── application.go   # Application struct and dependencies
+│   ├── context.go       # Context helper functions
+│   ├── errors.go        # Error response handlers
+│   └── helpers.go       # Helper functions (JSON, file processing)
+├── config/              # Configuration management
+│   └── config.go        # Configuration struct and parsing
+├── handlers/            # HTTP handlers organized by domain
+│   ├── healthcheck.go   # Health check handler
+│   └── handlers.go      # Placeholder handlers (to be moved from original files)
+├── middleware/          # HTTP middleware
+│   └── middleware.go    # Authentication, CORS, rate limiting, etc.
+├── routes/              # Route definitions
+│   └── routes.go        # All route configurations
+└── server/              # Server management
+    └── server.go        # HTTP server setup and graceful shutdown
+```
+
+## Benefits of New Structure
+
+### 1. **Separation of Concerns**
+- Each package has a single responsibility
+- Configuration is isolated from business logic
+- Middleware is separate from handlers
+- Routes are centralized and easy to manage
+
+### 2. **Better Maintainability**
+- Code is organized by function rather than being in one large package
+- Easy to locate specific functionality
+- Reduced coupling between components
+
+### 3. **Improved Testability**
+- Each package can be tested independently
+- Dependencies are clearly defined
+- Mock interfaces can be easily created
+
+### 4. **Scalability**
+- Easy to add new handlers, middleware, or configuration options
+- Clear patterns for extension
+- Better code reusability
+
+## Migration Steps
+
+### Phase 1: Core Structure (COMPLETED)
+- ✅ Created new folder structure
+- ✅ Moved configuration to `config/` package
+- ✅ Created `app/` package with Application struct
+- ✅ Moved helper functions to `app/helpers.go`
+- ✅ Moved error handlers to `app/errors.go`
+- ✅ Moved context functions to `app/context.go`
+- ✅ Created middleware package with proper function signatures
+- ✅ Created routes package with centralized route definitions
+- ✅ Created server package with graceful shutdown
+- ✅ Created new main.go entry point
+
+### Phase 2: Handler Migration (NEXT STEPS)
+- [ ] Move ideas handlers from `ideas.go` to `handlers/ideas.go`
+- [ ] Move user handlers from `users.go` to `handlers/users.go`
+- [ ] Move auth handlers from `tokens.go` to `handlers/auth.go`
+- [ ] Move OAuth handlers from `oauth.go` to `handlers/oauth.go`
+- [ ] Move profile handlers from `user_profiles.go` to `handlers/user_profiles.go`
+- [ ] Move file handlers from `files.go` to `handlers/files.go`
+
+### Phase 3: Testing and Cleanup
+- [ ] Test all endpoints with new structure
+- [ ] Remove old files after verification
+- [ ] Update any import paths in other parts of the application
+- [ ] Add unit tests for each package
+
+## Key Changes Made
+
+### 1. Configuration
+- Moved from `main.go` to dedicated `config/config.go`
+- Cleaner struct field names (uppercase for exported fields)
+- Centralized configuration loading
+
+### 2. Application Structure
+- Application struct moved to `app/application.go`
+- All application methods are now exported (TitleCase)
+- Clear dependency injection pattern
+
+### 3. Middleware
+- Converted from methods to functions returning middleware
+- Better composition and reusability
+- Type-safe middleware chaining
+
+### 4. Routes
+- Centralized route definitions
+- Clear middleware application
+- Easy to understand request flow
+
+### 5. Error Handling
+- Consistent error response patterns
+- Exported functions for reusability
+- Better separation of concerns
+
+## Usage Example
+
+The new structure allows for clean initialization:
+
+```go
+func main() {
+    // Load configuration
+    cfg, err := config.LoadConfig()
+    
+    // Initialize application
+    app := &app.Application{
+        Config: cfg,
+        // ... other dependencies
+    }
+    
+    // Start server with all middleware and routes
+    server.Serve(app)
+}
+```
+
+## Next Steps
+
+1. **Migrate Original Handlers**: Move all handler functions from the original files to the new `handlers/` package
+2. **Update Import Paths**: Ensure all imports point to the new structure
+3. **Test Thoroughly**: Verify all endpoints work correctly
+4. **Remove Old Files**: Clean up the original files once migration is complete
+5. **Add Documentation**: Document each package and its responsibilities
+
+This restructuring provides a solid foundation for a scalable, maintainable Go REST API following industry best practices.
