--- cmd/api/errors.go
@@ -1,98 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"net/http"
-)
-
-type ValidationError struct {
-	Errors map[string]string
-}
-
-func (app *application) logError(r *http.Request, err error) {
-	app.logger.PrintError(err, map[string]string{
-		"request_method": r.Method,
-		"request_url":    r.URL.String(),
-	})
-}
-
-func (app *application) errorResponse(w http.ResponseWriter, r *http.Request, status int, message any) {
-	env := envelope{"error": message}
-
-	err := app.writeJSON(w, status, env, nil)
-	if err != nil {
-		app.logError(r, err)
-		w.WriteHeader(500)
-	}
-}
-
-func (app *application) serverErrorResponse(w http.ResponseWriter, r *http.Request, err error) {
-	app.logError(r, err)
-	message := "the server encountered a problem and could not process your request"
-
-	app.errorResponse(w, r, http.StatusInternalServerError, message)
-}
-
-func (app *application) notFoundResponse(w http.ResponseWriter, r *http.Request) {
-	message := "the requested resource could not be found"
-	app.errorResponse(w, r, http.StatusNotFound, message)
-}
-
-func (app *application) methodNotAllowedResponse(w http.ResponseWriter, r *http.Request) {
-	message := fmt.Sprintf("the %s method is not supported for this resource", r.Method)
-	app.errorResponse(w, r, http.StatusMethodNotAllowed, message)
-}
-
-func (app *application) badRequestResponse(w http.ResponseWriter, r *http.Request, err error) {
-	app.errorResponse(w, r, http.StatusBadRequest, err.Error())
-}
-
-func (app *application) failedValidationResponse(w http.ResponseWriter, r *http.Request, errors map[string]string) {
-	app.errorResponse(w, r, http.StatusUnprocessableEntity, errors)
-}
-
-func (ve ValidationError) Error() string {
-	var errStr string
-	for field, msg := range ve.Errors {
-		errStr += fmt.Sprintf("%s: %s\n", field, msg)
-	}
-	return errStr
-}
-
-func (app *application) editConflictResponse(w http.ResponseWriter, r *http.Request) {
-	message := "unable to update the record due to an edit conflict, please try again"
-	app.errorResponse(w, r, http.StatusConflict, message)
-}
-
-func (app *application) rateLimitExceededResponse(w http.ResponseWriter, r *http.Request) {
-	message := "rate limit exceeded"
-	app.errorResponse(w, r, http.StatusTooManyRequests, message)
-}
-
-func (app *application) invalidCredentialsResponse(w http.ResponseWriter, r *http.Request) {
-	message := "invalid authentication credentials"
-	app.errorResponse(w, r, http.StatusUnauthorized, message)
-}
-
-func (app *application) invalidAuthenticationTokenResponse(w http.ResponseWriter, r *http.Request) {
-
-	w.Header().Set("WWW-Authenticate", "Bearer")
-
-	message := "invalid or missing authentication token"
-	app.errorResponse(w, r, http.StatusUnauthorized, message)
-}
-
-func (app *application) authenticationRequiredResponse(w http.ResponseWriter, r *http.Request) {
-	message := "you must be authenticated to access this resource"
-	app.errorResponse(w, r, http.StatusUnauthorized, message)
-}
-
-func (app *application) inactiveAccountResponse(w http.ResponseWriter, r *http.Request) {
-	message := "your user account must be activated before you can access this resource"
-	app.errorResponse(w, r, http.StatusForbidden, message)
-}
-
-func (app *application) notPermittedResponse(w http.ResponseWriter, r *http.Request) {
-	message := "your user account does not have the necessary permissions to access this resource"
-	app.errorResponse(w, r, http.StatusForbidden, message)
-}
