--- cmd/api/users.go
@@ -1,225 +1,0 @@
-package main
-
-import (
-	"errors"
-	"net/http"
-	"time"
-
-	"github.com/OpenConnectOUSL/backend-api-v1/internal/validator"
-
-	"github.com/OpenConnectOUSL/backend-api-v1/internal/data"
-)
-
-func (app *application) registerUserHandler(w http.ResponseWriter, r *http.Request) {
-	var input struct {
-		UserName string `json:"username"`
-		Email    string `json:"email"`
-		Password string `json:"password"`
-	}
-
-	err := app.readJSON(w, r, &input)
-	if err != nil {
-		app.badRequestResponse(w, r, err)
-		return
-	}
-
-	_, err = app.models.Users.GetByEmail(input.Email)
-	if err == nil {
-		app.failedValidationResponse(w, r, map[string]string{"email": "a user with this email address already exists"})
-		return
-	} else if !errors.Is(err, data.ErrRecordNotFound) {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	user := &data.User{
-		UserName:  input.UserName,
-		Email:     input.Email,
-		UserType:  "normal",
-		Activated: false,
-		HasProfileCreated: false,
-	}
-
-	err = user.Password.Set(input.Password)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	v := validator.New()
-
-	if data.ValidateUser(v, user); !v.Valid() {
-		app.failedValidationResponse(w, r, v.Errors)
-		return
-	}
-
-	err = app.models.Users.Insert(user)
-	if err != nil {
-		switch {
-		case errors.Is(err, data.ErrDuplicateEmail):
-			v.AddError("email", "a user with this email address already exists")
-			app.failedValidationResponse(w, r, v.Errors)
-		default:
-			app.serverErrorResponse(w, r, err)
-		}
-		return
-	}
-
-	err = app.models.Permissions.AddForUser(user.ID, "ideas:read")
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	token, err := app.models.Tokens.New(user.ID, 3*24*time.Hour, data.ScopeActivation)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	app.background(func() {
-
-		emailData := map[string]any{
-			"activationToken": token.Plaintext,
-			"userName":        user.UserName,
-			"frontendURL":     app.config.frontendURL,
-		}
-		err = app.mailer.Send(user.Email, "user_welcome", emailData)
-		if err != nil {
-			app.logger.PrintError(err, nil)
-		}
-	})
-
-	err = app.writeJSON(w, http.StatusAccepted, envelope{"user": user}, nil)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-	}
-
-}
-
-func (app *application) activateUserHandler(w http.ResponseWriter, r *http.Request) {
-	var input struct {
-		TokenPlainText string `json:"token"`
-	}
-
-	err := app.readJSON(w, r, &input)
-
-	if err != nil {
-		app.badRequestResponse(w, r, err)
-		return
-	}
-
-	v := validator.New()
-
-	if data.ValidateTokenPlaintext(v, input.TokenPlainText); !v.Valid() {
-		app.failedValidationResponse(w, r, v.Errors)
-		return
-	}
-
-	user, err := app.models.Users.GetForToken(data.ScopeActivation, input.TokenPlainText)
-	if err != nil {
-		switch {
-		case errors.Is(err, data.ErrRecordNotFound):
-			v.AddError("token", "invalid or expired activation token")
-			app.failedValidationResponse(w, r, v.Errors)
-		default:
-			app.serverErrorResponse(w, r, err)
-		}
-		return
-	}
-
-	user.Activated = true
-
-	err = app.models.Users.Update(user)
-	if err != nil {
-		switch {
-		case errors.Is(err, data.ErrEditConflict):
-			app.editConflictResponse(w, r)
-		default:
-			app.serverErrorResponse(w, r, err)
-		}
-		return
-	}
-
-	err = app.models.Permissions.AddForUser(user.ID, "ideas:write")
-
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	err = app.models.Tokens.DeleteAllForUser(data.ScopeActivation, user.ID)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	err = app.writeJSON(w, http.StatusOK, envelope{"user": user}, nil)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-	}
-}
-
-func (app *application) updateUserPasswordHandler(w http.ResponseWriter, r *http.Request) {
-	var input struct {
-		Password       string `json:"password"`
-		TokenPlaintext string `json:"token"`
-	}
-
-	err := app.readJSON(w, r, &input)
-	if err != nil {
-		app.badRequestResponse(w, r, err)
-		return
-	}
-
-	v := validator.New()
-
-	data.ValidatePasswordPlaintext(v, input.Password)
-	data.ValidateTokenPlaintext(v, input.TokenPlaintext)
-
-	if !v.Valid() {
-		app.failedValidationResponse(w, r, v.Errors)
-		return
-	}
-
-	user, err := app.models.Users.GetForToken(data.ScopePasswordReset, input.TokenPlaintext)
-	if err != nil {
-		switch {
-		case errors.Is(err, data.ErrEditConflict):
-			app.editConflictResponse(w, r)
-		default:
-			app.serverErrorResponse(w, r, err)
-		}
-		return
-	}
-
-	err = user.Password.Set(input.Password)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	err = app.models.Users.Update(user)
-	if err != nil {
-		switch {
-		case errors.Is(err, data.ErrEditConflict):
-			app.editConflictResponse(w, r)
-		default:
-			app.serverErrorResponse(w, r, err)
-		}
-		return
-	}
-	
-	err = app.models.Tokens.DeleteAllForUser(data.ScopePasswordReset, user.ID)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-		return
-	}
-
-	env := envelope{"message": "your password was successfully reset"}
-
-	err = app.writeJSON(w, http.StatusOK, env, nil)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-	}
-}
